рефакторинг CalcOrderAdditions и модификаторов
