+++ src/components/CalcOrderAdditions/AdditionsGroup.js
@@ -8,9 +8,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {ListItem, ListItemSecondaryAction, ListItemText} from 'material-ui/List';
-import IconButton from "material-ui/IconButton";
-import AddIcon from "material-ui-icons/AddCircleOutline";
-import RemoveIcon from "material-ui-icons/Delete";
+import IconButton from 'material-ui/IconButton';
+import AddIcon from 'material-ui-icons/AddCircleOutline';
+import RemoveIcon from 'material-ui-icons/Delete';
 import Divider from 'material-ui/Divider';
 import Collapse from 'material-ui/transitions/Collapse';
 import withStyles from './styles';
@@ -28,16 +28,16 @@ class AdditionsGroup extends React.Component {
     this.setState({
       count: this.state.count + 1,
     });
-  }
+  };
 
   handleRemove = () => {
     this.tabular && this.tabular.handleRemove();
-    if(this.state.count){
+    if(this.state.count) {
       this.setState({
         count: this.state.count - 1,
       });
     }
-  }
+  };
 
   render() {
 
@@ -52,8 +52,8 @@ class AdditionsGroup extends React.Component {
 
     return <div style={style}>
       <ListItem disableGutters className={classes.listitem}>
-        <IconButton title="Добавить строку" onClick={handleAdd}><AddIcon /></IconButton>
-        <IconButton title="Удалить строку" disabled={!count} onClick={handleRemove}><RemoveIcon /></IconButton>
+        <IconButton title="Добавить строку" onClick={handleAdd}><AddIcon/></IconButton>
+        <IconButton title="Удалить строку" disabled={!count} onClick={handleRemove}><RemoveIcon/></IconButton>
         <ListItemText primary={presentation}/>
         <ListItemSecondaryAction className={classes.secondary}>{count ? `${count} шт` : ''}</ListItemSecondaryAction>
       </ListItem>
@@ -80,6 +80,7 @@ AdditionsGroup.propTypes = {
   dp: PropTypes.object.isRequired,
   group: PropTypes.object.isRequired,
   scheme: PropTypes.object.isRequired,
+  count: PropTypes.number.isRequired,
   Renderer: PropTypes.func,
 };
